requirements.txt
fastapi~=0.115.11
pydantic~=2.10.6
langchain-core~=0.3.47
langgraph~=0.3.18
langchain-community~=0.3.20
langchain~=0.3.21
python-dotenv~=1.0.1
starlette~=0.46.1
httpx~=0.28.1
langchain-openai~=0.3.9